+++ packages/react/src/components/Accordion/AccordionContent.tsx
@@ -2,7 +2,7 @@ import cl from 'classnames';
 import React, { forwardRef, useContext } from 'react';
 
 import AnimateHeight from '../../utils/AnimateHeight';
-import { BodyLong } from '../Typography/BodyLong';
+import { BodyLong } from '../Typography/Body';
 
 import classes from './Accordion.module.css';
 import { AccordionItemContext } from './AccordionItem';
